+++ SnakeApp/app/src/main/java/com/hello/myapp/snakeapp/framework/Graphics.java
@@ -15,10 +15,10 @@ public interface Graphics {
 
     public void drawRect(int x, int y, int width, int height, int color);
 
-    public void drawPixmap(Pixmap pixmap, int x, int y, int srcX, int srcY,
+    public void drawPixmap(Pixmap pixmap, float x, float y, int srcX, int srcY,
                            int srcWidth, int srcHeight);
 
-    public void drawPixmap(Pixmap pixmap, int x, int y);
+    public void drawPixmap(Pixmap pixmap, float x, float y);
 
     public int getWidth();
