+++ SnakeApp/app/src/main/java/com/hello/myapp/snakeapp/framework/impl/AndroidGraphics.java
@@ -100,7 +100,7 @@ public class AndroidGraphics implements Graphics {
     }
 
     @Override
-    public void drawPixmap(Pixmap pixmap, int x, int y, int srcX, int srcY,
+    public void drawPixmap(Pixmap pixmap, float x, float y, int srcX, int srcY,
             int srcWidth, int srcHeight) {
         srcRect.left = srcX;
         srcRect.top = srcY;
